refactor(web): update import paths for reports and project components to use absolute imports
