+++ apps/web/core/components/reports/weekly-limit-report-export-menu.tsx
@@ -6,8 +6,8 @@ import { useTranslations } from 'next-intl';
 import { useMemo } from 'react';
 import { IOrganizationTeamList } from '@/core/types/interfaces';
 import { ITimeLimitReport, ITimeLimitReportByEmployee } from '@/core/types/interfaces/ITimeLimits';
-import { WeeklyLimitPDFDocument } from '../export-formats/pdf';
-import { WeeklyLimitByEmployeePDFDocument } from '../export-formats/pdf/grouped-by-employee';
+import { WeeklyLimitPDFDocument } from './export-formats/pdf';
+import { WeeklyLimitByEmployeePDFDocument } from './export-formats/pdf/grouped-by-employee';
 
 interface IProps {
 	data: ITimeLimitReport[];
@@ -54,7 +54,7 @@ export function WeeklyLimitExportMenu(props: IProps) {
 
 	return (
 		<Menu as="div" className="relative inline-block text-left">
-			<Menu.Button className=" w-full h-full items-center justify-between">
+			<Menu.Button className="items-center justify-between w-full h-full ">
 				<Button
 					type="button"
 					className=" border-gray-200 text-sm hover:bg-slate-100 min-w-fit text-black  h-[2.2rem] font-light hover:dark:bg-transparent"
@@ -76,7 +76,7 @@ export function WeeklyLimitExportMenu(props: IProps) {
 					static
 					className="absolute z-[999] left-1/2 -translate-x-1/2 mt-2 w-40 origin-top-right divide-y divide-gray-100 rounded-md bg-white dark:bg-dark-lighter shadow-lg ring-1 ring-black/5 focus:outline-none"
 				>
-					<div className="p-1 flex flex-col gap-1">
+					<div className="flex flex-col gap-1 p-1">
 						<Menu.Item>
 							{({ active }) => (
 								<button
